fix the double fault with a shitty RSP/RBP context switch.
This is the ugliest, worst piece of shit code I've ever written.

In the original version of this EFI implementation, we jump into the EFI
app with the Linux kernel's current page tables and pointers. RSP and
RBP point to a virtual address in high memory, i.e. something like
0xffffc90012345678. EFI applications in boot services expect all
pointers they are given to be 1:1 virt-to-phys mapped. The given RSP and
RBP obviously aren't.

Eventually, the Windows boot manager notices this and maps RSP/RBP to
their 1:1 location in the set of page tables it runs under. However,
when we return to Linux, and when Linux decides to switch to another set
of page tables internally, the stack pointer suddently points nowhere
and we double-page-fault.

So what do?

Upon jumping to the EFI app, find out where our _current stack_ is
physically located, add 1:1 mappings for the stack, and adjust RSP and
RBP to the 1:1 mapped location.

Upon coming back from the EFI app into Linux, reset the stack pointer to
the kernel-space virtual memory that is the "direct mapping of all
physical memory" at 0xffff888000000000.

See https://www.kernel.org/doc/Documentation/x86/x86_64/mm.txt

RSP before EFI and after EFI are not the same. The original Linux RSP
comes from the "vmalloc/ioremap space (vmalloc_base)" space, so the
original pointer is something like 0xffffc900......., and upon returning
we set it to 0xffff8880......

This just _happens to work_ for now. This is ugly. Do not do this.

What we actually need is a real context switch, and a stack just for the
EFI application. EFI apps expect 128KiB of stack space anyway, and a
Linux kernel stack is barely 2*PAGE_SIZE.

Signed-off-by: Chris Koch <chrisko@google.com>